Add files via upload
This agreement stipulates royalties ranging from 10% to 50% and levies a one-time fee of 2.5% for the first NFT produced on our platform. After that, there are no additional costs to mint and sell on the platform besides petrol. We don't provide free or in-bulk mints.   Although there is a 2.5 percent fee for the auction, we sell subscriptions starting at $4.95 per month instead. Giving you access to all the resources you need to promote your network, including blogs, forums, groups, and live streaming geographic mapping to get the lowest local prices on real goods. and much more.  You can sell the nfts it in dweb3 tokens or polygon mantic, and we utilize polygon to mint it. In order to trade the dweb3 tokens for money, we have an integrated exchange and 16 payment providers. Without email or any of the additional features that most websites include that force you to kyc before ever checking in with your wallet, we allow real web3 login utilizing the wallet login. We use an encrypted system that enables you to share your personal information with the admins for those who would like to upgrade. A code is generated that gives us a way to identify you with the code and stores the data off-server to protect it from scraps and malicious actors who might view your private information. The platform pays a fee ranging from 25% to 50% on every direct and personal invites made using the code we give you. In the white paper, all of this is mentioned. Never use your main wallet to login into the web3 connect use a dummy wallet that you can do purchase from and if you fall victim to the sweeper let us know we can assist you in removes the balances of assest left in your wallet. It takes maybe and hour and a little extra gas . It is not a hugh deal like they play it out to be